pyproject.toml
[project]
name = "naive-config-sync"
version = "0.1.0"
description = ""
authors = [
    {name = "unclefomotw",email = "unclefomotw@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "click (>=8.1.8,<9.0.0)"
]

[project.scripts]
naive-config-sync = "naive_config_sync.cli:main"

[tool.poetry]
packages = [{include = "naive_config_sync", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
